Changed output format of 'hostspool list' command
